orchestrator/requirements.txt
fastapi
uvicorn[standard]
redis
qdrant-client
pydantic
httpx
python-dotenv
pyyaml